+++ packages/terra-date-picker/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+3.15.0 - (May 7, 2019)
+------------------
+### Changed
+* Allow same day change to trigger onChange.
+
 3.14.0 - (April 24, 2019)
 ------------------
 ### Changed