+++ packages/terra-modal-manager/CHANGELOG.md
@@ -3,7 +3,10 @@ Changelog
 
 Unreleased
 ----------
-### Added 
+
+5.11.0 - (May 7, 2019)
+------------------
+### Added
 * Added new width 560px
 
 5.10.0 - (April 24, 2019)